amlapio v0.2.0 Amlapio

Adding a Map API to a GenServer’s state or module using an Agent to hold its state.

Amlapio can be use-d to generate “wrapper” functions that call Map functions on the state itself, or submaps held in the state.

All Map functions are supported except new.

Options

The options supported by Amlapio’s use are:

funs (optional)

The (optional) names of one or more functions supported by Map e.g. get, put, pop, etc.

If not supplied, wrappers are generated for all Map functions (except new).

namer (optional)

The namer is a function that is passed the map_name and fun_name (both Atoms) and is expected to return the name of the wrapper function (Atom).

For state wrappers, map_name will be nil.

agent

nil or one or more function names to generate Agent wrappers for.

nil means generate wrappers for the state itself.

The use below generates get and put wrappers for the buttons submap of an Agent

use Amlapio, agent: [:buttons], funs: [:get, :put]

(The example uses a utility function new_agent/1 in the repo’s tests to create the Agent process.)

This example shows how to get the the value of button1 key in the buttons submap of the Agent:

iex> new_agent(%{buttons: %{button1: :button1_value}})
...> |> buttons_get(:button1)
:button1_value

put works as expected:

iex> new_agent(%{buttons: %{}})
...> |> buttons_put(:button1, :button1_value)
...> |> Agent.get(fn s -> s end)
%{buttons: %{button1: :button1_value}}

The next use generates wrappers for all Map functions for the state of an Agent and names them e.g. agent_update

use Amlapio, agent: nil,
namer: fn _map_name, fun_name -> "agent_#{fun_name}" |> String.to_atom end

iex> new_agent(%{name: :ian, city: :london})
...> |> agent_update(:country, :uk, &(&1))
...> |> Agent.get(fn s -> s end)
%{name: :ian, city: :london, country: :uk}

iex> new_agent(%{name: :ian, city: :london})
...> |> agent_has_key?(:country)
false

behaviour_api or genserver_api

behaviour_api is an alias of genserver_api; either can be given.

Wrappers for a behaviour like GenServer have two functions. One is the client API function and the other the callback function (e.g. handle_call/3). The callback has to be defined together with all the other callbacks of the same name (else the compiler will warn).

behaviour_api generates the client wrapper(s), here for the GenServer state and call them e.g. gen_state_get:

use Amlapio, genserver_api: nil,
namer: fn _map_name, fun_name -> "gen_state_#{fun_name}" |> String.to_atom end

behaviour_callback or genserver_handle_call

behaviour_callback is an alias of genserver_handle_call; either can be given.

behaviour_callback generates the callback function for the client wrappers:

use Amlapio, genserver_handle_call: nil,
namer: fn _map_name, fun_name -> "gen_state_#{fun_name}" |> String.to_atom end

With both client and callback generated, the wrappers can be called:

iex> new_genserver(%{name: :ian, city: :london})
...> |> gen_state_update(:country, :uk, &(&1))
...> |> gen_state_get(:country)
:uk

behaviour_module

The default behaviour_module is GenServer but other behaviours that support call and handle_call (e.g. GenMQTT) can be given.

Here is the GenServer example above, reworked for GenMQTT

use Amlapio, behaviour_module: GenMQTT, behaviour_api: nil,
namer: fn _map_name, fun_name -> "mqtt_#{fun_name}" |> String.to_atom end

use Amlapio, behaviour_callback: nil,
namer: fn _map_name, fun_name -> "mqtt_#{fun_name}" |> String.to_atom end

iex> new_mqtt(%{host: "localhost"})
...> |> mqtt_put(:port, 1883)
...> |> mqtt_get(:port)
1883

Other Examples

The README has similar examples as does the tests in the repo.

Types

map_name()

map_name() :: nil | atom

The name of the state or submap; nil => state

maybe_quoted_behaviour_module()

maybe_quoted_behaviour_module() :: module | Macro.t

The maybe quoted name of the module supplying the call function e.g. GenServer

maybe_quoted_namer_fun()

maybe_quoted_namer_fun() :: Macro.t | (map_name, name -> name)

Maybe quoted namer fun

name()

name() :: atom

names()

names() :: nil | name | [name]

The names of the Map functions to generate wrappers for

use_option()

use_option ::
  {:agent, names} |
  {:genserver_api, names} |
  {:genserver_handle_call, names} |
  {:behaviour_api, names} |
  {:behaviour_callback, names} |
  {:behaviour_function, name} |
  {:behaviour_module, maybe_quoted_behaviour_module} |
  {:namer, maybe_quoted_namer_fun}

use_options()

use_options() :: [use_option]

The supported options passed to the use call.

Macros

__using__(opts)

__using__(term, use_options) :: [Macro.t]

See above for the supported options to the use call.